Save preference to show system apps
